Create Code to sort and print out list in alphanumeric order
Because I used a Binary Search Tree in this program it sorted the items as they were added so by printing the search tree in order the code prints the courses in alphanumeric order. I have included the code to add the item into the binary search tree to show when it is sorted as well as the code to print this sorted list to fully show how the code accomplishes this task.